Added Upload Image to EditHousingPage, Fixed Bugs
Added Upload Image to EditHousingPage.
Fixed the previous FlatList issue such that it was a stupid mistake. Everything is working as expected.
Fixed the HouseFavButton such that when it is loaded, the favorite information will be up to date.
Fixed the House model such that the picture starts with an empty array, instead of an array with an empty string.
Fixed the HousingSearchPage to display the housing in post_date's descending order.